Fix possible ProgressCounter bug
ProgressCounter’s init method now takes a collection instead of an Int.
It generates the number it needs itself via len(), which always returns
a positive number or zero. This avoids a potential bug wherein,
when len() is later called on the str representation of that number,
a negative Int would have its length misreported. Better to avoid the
possibility of negative Ints altogether.