Make cache funcs accept and return datetime objs
The refactoring for this exposed some buggy areas around how
the cache_details named tuple was implemented. To fix this
get_checksum_input_values()'s return tuple can now return None for
cache_details, and the default for functions that accept cache_details
is now None. The documentation now assumes that, if cache_details is 
present, none of its values are None.